AI_Code_Reviewer_LLM/requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]

